Create README.md
This is the a reference to https://bitbucket.org/adigsd/backend-flitetrakr.git challenge.
The application does not need any dependencies to run. There is only a resource file input.txt that includes some commands to test the application.
The app is build on intelij and can be run by generating a jar file from it.